refactor : 카테고리 요청/반환 json의 키 형식 동일 - id -> categoryId - name -> name - icon -> icon - priority -> priority - isVisible -> isVisible - type -> type
Related to : #1
