+++ src/main/java/com/umc5th/muffler/domain/category/dto/NewCategoryRequest.java
@@ -11,7 +11,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class NewCategoryRequest {
     @NotBlank
-    private String categoryName;
+    private String name;
     @NotBlank
     private String icon;
     @NotNull
