+++ src/main/java/com/umc5th/muffler/domain/category/dto/NewCategoryResponse.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class NewCategoryResponse {
-    private Long id;
+    private Long categoryId;
 }
