+++ src/main/java/com/umc5th/muffler/domain/category/dto/OutlineCategoryDTO.java
@@ -9,6 +9,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class OutlineCategoryDTO {
-    private Long id;
+    private Long categoryId;
     private String name;
 }
