+++ src/main/java/com/umc5th/muffler/domain/category/dto/UpdateCategoryNameIconRequest.java
@@ -13,7 +13,7 @@ public class UpdateCategoryNameIconRequest {
     @Positive
     private Long categoryId;
     @NotBlank
-    private String categoryName;
+    private String name;
     @NotBlank
-    private String categoryIcon;
+    private String icon;
 }
